fix: change index.ts module file #88
